sparse_is_na {sparsevctrs}	R Documentation

Detect Pressence of Missing Values

Description

Detect Pressence of Missing Values

Usage

sparse_is_na(x, type = "logical")

Arguments

x
A sparse vector.

type
A single string. Most be one of logical or integer. Determines the resulting vector. If type = integer then a sparse vector is returned.

Details

This function, as with any of the other helper functions assumes that the input x is a sparse numeric vector. This is done for performance reasons, and it is thus the users responsibility to perform input checking.

Note that the resulting vector will be not be a sparse vector.

Value

A logical vector or sparse integer vector.

See Also

sparse_which_na()

Examples

sparse_is_na(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000)
)

sparse_is_na(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000),
  type = "integer"
)

[Package sparsevctrs version 0.3.6 Index]