sparse_replace_na {sparsevctrs}	R Documentation

Replace NAs with specified values in sparse vectors

Description

Replace NAs with specified values in sparse vectors

Usage

sparse_replace_na(x, replace)

Arguments

x
A sparse vector.

replace
A single value.

Details

This function, as with any of the other helper functions assumes that the input x is a sparse numeric vector. This is done for performance reasons, and it is thus the users responsibility to perform input checking. The replace is likewise not type or length checked.

The output type will match the values after coercion happens during replacement.

Value

A sparse vector.

Examples

sparse_replace_na(
  sparse_double(c(10, NA, 11), c(1, 5, 10), 10),
  5
)

sparse_replace_na(
  sparse_integer(c(10L, NA, 11L), c(1, 5, 10), 10),
  5L
)

sparse_replace_na(
  sparse_character(c("A", NA, "E"), c(2, 5, 10), 10),
  "missing"
)

[Package sparsevctrs version 0.3.6 Index]