sparse_sd {sparsevctrs}	R Documentation

Calculate standard diviation from sparse vectors

Description

Calculate standard diviation from sparse vectors

Usage

sparse_sd(x, na_rm = FALSE)

Arguments

x
A sparse numeric vector.

na_rm
Logical, whether to remove missing values. Defaults to FALSE.

Details

This function, as with any of the other helper functions assumes that the input x is a sparse numeric vector. This is done for performance reasons, and it is thus the users responsibility to perform input checking.

Much like sd() it uses the denominator n-1.

Value

single numeric value.

Examples

sparse_sd(
  sparse_double(1000, 1, 1000)
)

sparse_sd(
  sparse_double(1000, 1, 1000, default = 1)
)

sparse_sd(
  sparse_double(c(10, 50, 11), c(1, 50, 111), 1000)
)

sparse_sd(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000)
)

sparse_sd(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000),
  na_rm = TRUE
)

[Package sparsevctrs version 0.3.6 Index]