sparse_which_na {sparsevctrs}	R Documentation

Which indices are Missing Values

Description

Which indices are Missing Values

Usage

sparse_which_na(x)

Arguments

x
A sparse vector.

Details

This function, as with any of the other helper functions assumes that the input x is a sparse numeric vector. This is done for performance reasons, and it is thus the users responsibility to perform input checking.

Value

A logical vector.

See Also

sparse_is_na()

Examples

sparse_which_na(
  sparse_double(c(10, NA, 11), c(1, 50, 111), 1000)
)

[Package sparsevctrs version 0.3.6 Index]