sparsevctrs-package {sparsevctrs}	R Documentation

sparsevctrs: Sparse Vectors for Use in Data Frames

Description

Provides sparse vectors powered by ALTREP (Alternative Representations for R Objects) that behave like regular vectors, and can thus be used in data frames. Also provides tools to convert between sparse matrices and data frames with sparse columns and functions to interact with sparse vectors.

Author(s)

Maintainer: Emil Hvitfeldt emil.hvitfeldt@posit.co (ORCID)

Other contributors:

Davis Vaughan davis@posit.co [contributor]

Posit Software, PBC (ROR) [copyright holder, funder]

See Also

Useful links:

https://github.com/r-lib/sparsevctrs

https://r-lib.github.io/sparsevctrs/

Report bugs at https://github.com/r-lib/sparsevctrs/issues

[Package sparsevctrs version 0.3.6 Index]